backend/pyproject.toml
[tool.poetry]
name = "rag-chat"
version = "0.2.0"
description = "A RAG-based Chatbot"
authors = ["Jianshao"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-core = "^0.12.5"
llama-index-readers-file = "^0.4.1"
llama-index-vector-stores-chroma = "^0.4.1"
llama-index-embeddings-openai = "^0.3.1"
llama-index-embeddings-gemini = "^0.3.0"
llama-index-embeddings-ollama = "^0.5.0"
llama-index-llms-openai = "^0.3.4"
llama-index-llms-gemini = "^0.4.0"
llama-index-llms-ollama = "^0.5.0"
llama-index-agent-openai = "^0.4.1"
fastapi = "^0.115.6"
aiostream = "^0.6.4"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
